try propagating a bunch of get_gpu_memory* int32->size_t changes through HSA & HIP

+++ lib/hsa/hsaAsyncCopyGain.hpp
@@ -27,7 +27,7 @@ public:
 
 private:
     struct Buffer* gain_buf;
-    int32_t gain_len;
+    size_t gain_len;
     int32_t gain_buf_id;
     int32_t gain_buf_finalize_id;
     int32_t gain_buf_precondition_id;
